requirements.txt
# Core
python-dotenv>=1.0
pydantic-settings>=2.0
pyyaml>=6.0
pandas>=2.0
pandera>=0.18
scikit-learn>=1.4
numpy>=1.26
joblib>=1.3
fastapi>=0.109
uvicorn>=0.27

# Dev
pytest>=7.0
pytest-cov>=4.0
ruff>=0.1
pre-commit>=3.0
httpx>=0.26
